refactor: remove original trpc client, migrate trpc-rxjs-app to trpc-app
