+++ apps/trpc-app/index.html
@@ -1,15 +1,15 @@
 <!DOCTYPE html>
-<html lang='en'>
-<head>
-  <meta charset='utf-8' />
-  <title>SPARTAN - Notes App</title>
-  <base href='/' />
-  <meta name='viewport' content='width=device-width, initial-scale=1' />
-  <link rel="icon" type="image/x-icon" href="/favicon.ico">
-  <link rel='stylesheet' href='/src/styles.css' />
-</head>
-<body>
-<trpc-app-root></trpc-app-root>
-<script type='module' src='/src/main.ts'></script>
-</body>
+<html class="dark h-full" lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <title>SPARTAN - Notes App</title>
+    <base href="/" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <link rel="icon" type="image/x-icon" href="/favicon.ico" />
+    <link rel="stylesheet" href="/src/styles.css" />
+  </head>
+  <body class="h-full bg-zinc-900">
+    <trpc-app-root></trpc-app-root>
+    <script type="module" src="/src/main.ts"></script>
+  </body>
 </html>
