+++ apps/trpc-app/src/server/trpc/context.ts
@@ -1,4 +1,5 @@
-import { inferAsyncReturnType } from '@trpc/server';
+import { inferAsyncReturnType } from "@trpc/server";
+
 /**
  * Creates context for an incoming request
  * @link https://trpc.io/docs/context
