--- apps/trpc-rxjs-app/project.json
@@ -1,71 +1,0 @@
-{
-  "name": "trpc-rxjs-app",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "projectType": "application",
-  "sourceRoot": "apps/trpc-rxjs-app/src",
-  "targets": {
-    "build": {
-      "dependsOn": ["^build"],
-      "executor": "@nx/vite:build",
-      "outputs": [
-        "{options.outputPath}",
-        "dist/apps/trpc-rxjs-app/.nitro",
-        "dist/apps/trpc-rxjs-app/ssr",
-        "dist/apps/trpc-rxjs-app/analog"
-      ],
-      "options": {
-        "configFile": "apps/trpc-rxjs-app/vite.config.ts",
-        "outputPath": "dist/apps/trpc-rxjs-app/client"
-      },
-      "defaultConfiguration": "production",
-      "configurations": {
-        "development": {
-          "mode": "development"
-        },
-        "production": {
-          "sourcemap": false,
-          "mode": "production"
-        }
-      }
-    },
-    "serve": {
-      "executor": "@nx/vite:dev-server",
-      "defaultConfiguration": "development",
-      "options": {
-        "buildTarget": "trpc-rxjs-app:build",
-        "port": 4215
-      },
-      "configurations": {
-        "development": {
-          "buildTarget": "trpc-rxjs-app:build:development",
-          "hmr": true
-        },
-        "production": {
-          "buildTarget": "trpc-rxjs-app:build:production"
-        }
-      }
-    },
-    "extract-i18n": {
-      "executor": "@angular-devkit/build-angular:extract-i18n",
-      "options": {
-        "browserTarget": "trpc-rxjs-app:build"
-      }
-    },
-    "lint": {
-      "executor": "@nx/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": [
-          "apps/trpc-rxjs-app/**/*.ts",
-          "apps/trpc-rxjs-app/**/*.html"
-        ]
-      }
-    },
-    "test": {
-      "executor": "@nx/vite:test",
-      "outputs": ["apps/trpc-rxjs-app/coverage"]
-    }
-  },
-  "tags": [],
-  "implicitDependencies": ["trpc"]
-}
