--- apps/trpc-rxjs-app/src/app.config.server.ts
@@ -1,10 +1,0 @@
-import { ApplicationConfig, mergeApplicationConfig } from "@angular/core";
-import { provideServerRendering } from "@angular/platform-server";
-import { appConfig } from "./app.config";
-import { provideClientHydration } from "@angular/platform-browser";
-
-const serverConfig: ApplicationConfig = {
-  providers: [provideServerRendering(), provideClientHydration()],
-};
-
-export const config = mergeApplicationConfig(appConfig, serverConfig);
