--- apps/trpc-rxjs-app/src/app.config.ts
@@ -1,12 +1,0 @@
-import { ApplicationConfig } from "@angular/core";
-import { provideClientHydration } from "@angular/platform-browser";
-import { provideFileRouter } from "@analogjs/router";
-import { provideTRPCClient } from "./trpc-client";
-
-export const appConfig: ApplicationConfig = {
-  providers: [
-    provideFileRouter(),
-    provideClientHydration(),
-    provideTRPCClient(),
-  ],
-};
