--- apps/trpc-rxjs-app/src/main.server.ts
@@ -1,20 +1,0 @@
-import "zone.js/node";
-import { enableProdMode } from "@angular/core";
-import { renderApplication } from "@angular/platform-server";
-import { AppComponent } from "./app/app.component";
-import { bootstrapApplication } from "@angular/platform-browser";
-import { config } from "./app.config.server";
-
-if (import.meta.env.PROD) {
-  enableProdMode();
-}
-
-const bootstrap = () => bootstrapApplication(AppComponent, config);
-
-export default async function render(url: string, document: string) {
-  const html = await renderApplication(bootstrap, {
-    document,
-    url,
-  });
-  return html;
-}
