--- apps/trpc-rxjs-app/src/main.ts
@@ -1,9 +1,0 @@
-import "zone.js";
-import { bootstrapApplication } from "@angular/platform-browser";
-
-import { AppComponent } from "./app/app.component";
-import { appConfig } from "./app.config";
-
-bootstrapApplication(AppComponent, appConfig).catch((err) =>
-  console.error(err)
-);
