--- apps/trpc-rxjs-app/src/note.ts
@@ -1,5 +1,0 @@
-export type Note = {
-  id: number;
-  note: string;
-  createdAt: Date;
-};
