--- apps/trpc-rxjs-app/src/server/routes/health.ts
@@ -1,3 +1,0 @@
-import { eventHandler } from "h3";
-
-export default eventHandler(() => ({ success: true }));
