--- apps/trpc-rxjs-app/src/server/routes/trpc/[trpc].ts
@@ -1,8 +1,0 @@
-import { appRouter } from "../../trpc/routers";
-import { createContext } from "../../trpc/context";
-import { createTrpcNitroHandler } from "@analogjs/trpc";
-// export API handler
-export default createTrpcNitroHandler({
-  router: appRouter,
-  createContext,
-});
