--- apps/trpc-rxjs-app/src/server/trpc/context.ts
@@ -1,8 +1,0 @@
-import { inferAsyncReturnType } from "@trpc/server";
-
-/**
- * Creates context for an incoming request
- * @link https://trpc.io/docs/context
- */
-export const createContext = () => ({});
-export type Context = inferAsyncReturnType<typeof createContext>;
