--- apps/trpc-rxjs-app/src/server/trpc/routers/index.ts
@@ -1,8 +1,0 @@
-import { router } from "../trpc";
-import { noteRouter } from "./notes";
-
-export const appRouter = router({
-  note: noteRouter,
-});
-// export type definition of API
-export type AppRouter = typeof appRouter;
