--- apps/trpc-rxjs-app/src/server/trpc/trpc.ts
@@ -1,13 +1,0 @@
-import { initTRPC } from "@trpc/server";
-import { Context } from "./context";
-import superjson from "superjson";
-
-const t = initTRPC.context<Context>().create({
-  transformer: superjson,
-});
-/**
- * Unprotected procedure
- **/
-export const publicProcedure = t.procedure;
-export const router = t.router;
-export const middleware = t.middleware;
