--- apps/trpc-rxjs-app/src/test-setup.ts
@@ -1,9 +1,0 @@
-import "@analogjs/vite-plugin-angular/setup-vitest";
-
-import { BrowserDynamicTestingModule, platformBrowserDynamicTesting } from "@angular/platform-browser-dynamic/testing";
-import { getTestBed } from "@angular/core/testing";
-
-getTestBed().initTestEnvironment(
-  BrowserDynamicTestingModule,
-  platformBrowserDynamicTesting()
-);
