--- apps/trpc-rxjs-app/tsconfig.app.json
@@ -1,16 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "../../dist/out-tsc",
-    "types": [],
-    "target": "ES2022",
-    "useDefineForClassFields": false
-  },
-  "files": ["src/main.ts", "src/main.server.ts"],
-  "include": [
-    "src/**/*.d.ts",
-    "src/app/routes/**/*.ts",
-    "src/app/pages/**/*.page.ts"
-  ],
-  "exclude": ["**/*.test.ts", "**/*.spec.ts"]
-}
