--- apps/trpc-rxjs-app/tsconfig.editor.json
@@ -1,7 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "include": ["**/*.ts"],
-  "compilerOptions": {
-    "types": ["node", "vitest/globals"]
-  }
-}
