--- apps/trpc-rxjs-app/tsconfig.spec.json
@@ -1,9 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "../../dist/out-tsc",
-    "types": ["node", "vitest/globals"]
-  },
-  "files": ["src/test-setup.ts", "src/polyfills.ts"],
-  "include": ["src/**/*.spec.ts", "**/*.d.ts"]
-}
