+++ packages/trpc/src/index.ts
@@ -1,5 +1,3 @@
 export * from './lib/client/client';
-export * from './lib/client/rxjs/client';
 export * from './lib/server/server';
 export * from './lib/utils/wait-for';
-export * from './lib/client/client-shared';
