--- packages/trpc/src/lib/client/client-shared.ts
@@ -1,7 +1,0 @@
-import { AnyRouter } from "@trpc/server";
-import { CreateTRPCClientOptions } from "@trpc/client/src/createTRPCUntypedClient";
-
-export type TrpcOptions<T extends AnyRouter> = {
-  url: string;
-  options: Partial<CreateTRPCClientOptions<T>>;
-};
