+++ packages/trpc/src/lib/client/client.ts
@@ -1,19 +1,20 @@
-/**
- * Inspired by this awesome project to integrate trpc more into the angular way
- * of doing things https://github.com/Dafnik/ngx-trpc
- */
 import { InjectionToken, Provider, TransferState } from "@angular/core";
 import "isomorphic-fetch";
-import { createTRPCProxyClient, httpBatchLink } from "@trpc/client";
+import { httpBatchLink } from "@trpc/client";
 import { AnyRouter } from "@trpc/server";
 import { transferStateLink } from "./links/transfer-state-link";
 import { provideTrpcCacheState, provideTrpcCacheStateStatusManager, tRPC_CACHE_STATE } from "./cache-state";
-import { TrpcOptions } from "./client-shared";
+import { createTRPCRxJSProxyClient } from "./trpc-rxjs-proxy";
+import { CreateTRPCClientOptions } from "@trpc/client/src/createTRPCUntypedClient";
+
+export type TrpcOptions<T extends AnyRouter> = {
+  url: string;
+  options: Partial<CreateTRPCClientOptions<T>>;
+};
 
 export type TrpcClient<AppRouter extends AnyRouter> = ReturnType<
-  typeof createTRPCProxyClient<AppRouter>
+  typeof createTRPCRxJSProxyClient<AppRouter>
 >;
-
 const tRPC_INJECTION_TOKEN = new InjectionToken<unknown>(
   '@analogjs/trpc proxy client'
 );
@@ -29,7 +30,7 @@ export const createTrpcClient = <AppRouter extends AnyRouter>({
       useFactory: () => {
         // eslint-disable-next-line @typescript-eslint/ban-ts-comment
         // @ts-ignore TODO: figure out why TS is complaining
-        return createTRPCProxyClient<AppRouter>({
+        return createTRPCRxJSProxyClient<AppRouter>({
           transformer: options?.transformer,
           links: [
             ...(options?.links ?? []),
