+++ packages/trpc/src/lib/client/trpc-rxjs-proxy.ts
@@ -5,33 +5,27 @@ import type {
   AnyRouter,
   CombinedDataTransformer,
   DataTransformerOptions,
-  DefaultDataTransformer,
   inferProcedureInput,
   inferProcedureOutput,
+  IntersectionError,
   ProcedureArgs,
   ProcedureRouterRecord,
   ProcedureType
 } from "@trpc/server";
-import { ClientDataTransformerOptions } from "@trpc/server";
 import { createFlatProxy, createRecursiveProxy, inferTransformedProcedureOutput } from "@trpc/server/shared";
-import { inferObservableValue, observable, share } from "@trpc/server/observable";
-import {
-  Operation,
-  OperationContext,
-  OperationLink,
-  OperationResultObservable,
-  TRPCClientRuntime
-} from "@trpc/client/src/links/types";
-import { Observable as TrpcObservable } from "@trpc/server/src/observable/types";
+import { inferObservableValue, share } from "@trpc/server/observable";
+import { OperationContext, OperationLink, TRPCClientRuntime } from "@trpc/client/src/links/types";
 import { Observable as RxJSObservable } from "rxjs";
-import { TRPCClientError, TRPCLink } from "@trpc/client";
+import { TRPCClientError, TRPCRequestOptions } from "@trpc/client";
+import { createChain, CreateTRPCClientOptions, TRPCType } from "./shared-internal";
+import { Observable as TrpcObservable } from "@trpc/server/src/observable/types";
 
 // Changed to rxjs observable
 type Resolver<TProcedure extends AnyProcedure> = (
   ...args: ProcedureArgs<TProcedure['_def']>
 ) => RxJSObservable<inferTransformedProcedureOutput<TProcedure>>;
 
-// Removed subscription
+// Removed subscription and using new type
 type DecorateProcedure<
   TProcedure extends AnyProcedure,
   TRouter extends AnyRouter
@@ -45,7 +39,7 @@ type DecorateProcedure<
     }
   : never;
 
-// Removed subscription
+// Removed subscription and using new type
 type DecoratedProcedureRecord<
   TProcedures extends ProcedureRouterRecord,
   TRouter extends AnyRouter
@@ -60,7 +54,7 @@ type DecoratedProcedureRecord<
     : never;
 };
 
-// Removed subscription
+// Removed subscription and using new type
 const clientCallTypeMap: Record<
   keyof DecorateProcedure<any, any>,
   ProcedureType
@@ -78,7 +72,7 @@ type UntypedClientProperties =
   | 'query'
   | 'mutation';
 
-// Nothing changed, only using different types
+// Nothing changed, only using new types
 export type CreateTRPCProxyClient<TRouter extends AnyRouter> =
   DecoratedProcedureRecord<
     TRouter['_def']['record'],
@@ -89,9 +83,7 @@ export type CreateTRPCProxyClient<TRouter extends AnyRouter> =
       : IntersectionError<UntypedClientProperties & keyof TProcedureRecord>
     : never;
 
-/**
- * Nothing changed, only using different types
- */
+// Nothing changed, only using new types
 function createTRPCRxJSClientProxy<TRouter extends AnyRouter>(
   client: TRPCClient<TRouter>
 ) {
@@ -124,8 +116,6 @@ export function createTRPCRxJSProxyClient<TRouter extends AnyRouter>(
   const proxy = createTRPCRxJSClientProxy(client as TRPCClient<TRouter>);
   return proxy;
 }
-
-type TRPCType = 'query' | 'mutation';
 
 /**
  * Removed subscription method;
@@ -250,92 +240,3 @@ function trpcObservableToRxJsObservable<TValue>(
     };
   });
 }
-
-/*
- * One to one copy of the trpc client internal code
- * Nothing was changed, but we can not import these methods because
- * they are not exported
- */
-
-type IntersectionError<TKey extends string> =
-  `The property '${TKey}' in your router collides with a built-in method, rename this router or procedure on your backend.`;
-
-export interface TRPCRequestOptions {
-  /**
-   * Pass additional context to links
-   */
-  context?: OperationContext;
-  signal?: AbortSignal;
-}
-
-function createChain<
-  TRouter extends AnyRouter,
-  TInput = unknown,
-  TOutput = unknown
->(opts: {
-  links: OperationLink<TRouter, TInput, TOutput>[];
-  op: Operation<TInput>;
-}): OperationResultObservable<TRouter, TOutput> {
-  return observable((observer) => {
-    function execute(index = 0, op = opts.op) {
-      const next = opts.links[index];
-      if (!next) {
-        throw new Error(
-          'No more links to execute - did you forget to add an ending link?'
-        );
-      }
-      const subscription = next({
-        op,
-        next(nextOp) {
-          const nextObserver = execute(index + 1, nextOp);
-
-          return nextObserver;
-        },
-      });
-      return subscription;
-    }
-
-    const obs$ = execute();
-    return obs$.subscribe(observer);
-  });
-}
-
-type CreateTRPCClientOptions<TRouter extends AnyRouter> =
-  | CreateTRPCClientBaseOptions<TRouter> & {
-      links: TRPCLink<TRouter>[];
-    };
-
-type CreateTRPCClientBaseOptions<TRouter extends AnyRouter> =
-  TRouter['_def']['_config']['transformer'] extends DefaultDataTransformer
-    ? {
-        /**
-         * Data transformer
-         *
-         * You must use the same transformer on the backend and frontend
-         * @link https://trpc.io/docs/data-transformers
-         **/
-        transformer?: 'You must set a transformer on the backend router';
-      }
-    : TRouter['_def']['_config']['transformer'] extends DataTransformerOptions
-    ? {
-        /**
-         * Data transformer
-         *
-         * You must use the same transformer on the backend and frontend
-         * @link https://trpc.io/docs/data-transformers
-         **/
-        transformer: TRouter['_def']['_config']['transformer'] extends CombinedDataTransformer
-          ? DataTransformerOptions
-          : TRouter['_def']['_config']['transformer'];
-      }
-    : {
-        /**
-         * Data transformer
-         *
-         * You must use the same transformer on the backend and frontend
-         * @link https://trpc.io/docs/data-transformers
-         **/
-        transformer?:
-          | /** @deprecated **/ ClientDataTransformerOptions
-          | CombinedDataTransformer;
-      };
